+++ lib/testing_api_onlyoffice_com/test_instance/main_page/workspace/workspace_faq_page.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 module TestingApiOnlyOfficeCom
-  # https://user-images.githubusercontent.com/668524/146583286-eddeaf82-a560-4e56-a79c-d87c1ccf043c.png
-  # Class for `https://api.onlyoffice.com/portals/faq` page
-  class CommunityServerFaqPage
+  # https://github.com/ONLYOFFICE-QA/testing-api.onlyoffice.com/assets/60688343/28a0e03f-5001-40e9-9b49-e3892e7d19cf
+  # Class for `https://api.teamlab.info/portals/workspaceapi/faq`
+  class WorkspaceFaqPage
     def initialize(instance)
       @instance = instance
       wait_to_load