+++ spec/testing-api.onlyoffice.com/workspace/documentation_links_spec.rb
@@ -1,17 +1,17 @@
 # frozen_string_literal: true
 
 require 'spec_helper'
-describe 'community_server_documentation' do
-  test_manager = TestingApiOnlyOfficeCom::TestManager.new(suite_name: 'Community Server Documentation Links', plan_name: config.to_s)
+describe 'workspace_documentation' do
+  test_manager = TestingApiOnlyOfficeCom::TestManager.new(suite_name: 'Workspace Documentation Links', plan_name: config.to_s)
 
   before do
     @instance = TestingApiOnlyOfficeCom::TestInstance.new(config)
     @api_page = @instance.go_to_main_page
   end
 
-  describe 'document_builder_documentation_links' do
+  describe 'workspace_documentation_links' do
     before do
-      @api_community_server_page = @api_page.go_to_workspace_api
+      @api_workspace_page = @api_page.go_to_workspace_api
     end
 
     after do |example|
@@ -20,7 +20,7 @@ describe 'community_server_documentation' do
     end
 
     it 'check all links in documentation are visible' do
-      result, failed = @api_community_server_page.community_server_links_ok?
+      result, failed = @api_workspace_page.workspace_links_ok?
       expect(result).to be_truthy, "Page #{@instance.webdriver.driver.current_url}\n\nNot found api links: #{failed}"
     end
   end